Permissions: convert hook to util
