+++ ui/components/app/permissions-connect-permission-list/permissions-connect-permission-list.js
@@ -1,14 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { usePermissionDescriptions } from '../../../hooks/usePermissionDescriptions';
+import { getPermissionDescription } from '../../../helpers/utils/permission';
+import { useI18nContext } from '../../../hooks/useI18nContext';
 
 export default function PermissionsConnectPermissionList({ permissions }) {
-  const getPermissionDescription = usePermissionDescriptions();
+  const t = useI18nContext();
 
   return (
     <div className="permissions-connect-permission-list">
       {Object.keys(permissions).map((permission) => {
         const { label, leftIcon, rightIcon } = getPermissionDescription(
+          t,
           permission,
         );
